refactor: 자동차 이름 조건 검사 함수 분리
