+++ src/index.js
@@ -1,4 +1,5 @@
 import { CARRACING } from './constants.js';
+import checkCarName from './checkInput/checkCarName.js';
 const carNameInput = document.getElementById('car-names-input');
 const carNameSubmit = document.getElementById('car-names-submit');
 const countInput = document.getElementById('racing-count-input');
@@ -11,48 +12,6 @@ export default function racingGame() {
 
 function Car(name) {
   this.name = name;
-}
-
-function checkCarNameDuplicate(carNameList) {
-  for (let name of carNameList) {
-    let count = carNameList.filter(element => name === element).length;
-
-    if (count >= 2) {
-      return false;
-    }
-  }
-
-  return true;
-}
-
-function checkCarNameLen(carNameList) {
-  for (let name of carNameList) {
-    if (name.length >= 6) {
-      return false;
-    }
-  }
-
-  return true;
-}
-
-function checkCarNameComma(carNameList) {
-  if (carNameList.length === 1) {
-    return false;
-  }
-
-  return true;
-}
-
-function checkCarName(carNameList) {
-  if (
-    checkCarNameDuplicate(carNameList) &&
-    checkCarNameLen(carNameList) &&
-    checkCarNameComma(carNameList)
-  ) {
-    return true;
-  }
-
-  return false;
 }
 
 function getCarName() {
